pyproject.toml
[project]
name = "patient-simulation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "matplotlib>=3.10.8",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
]
